requirements.txt
flask==3.0.0
flask-cors==4.0.0
psycopg2==2.9.9
bcrypt==4.0.1
python-dotenv==1.0.0
stripe==7.6.0
tiktoken==0.5.1
itsdangerous==2.1.2
schedule==1.2.0